refactor: console, web을 기준으로 패키지 구조 변경
