+++ src/main/java/roomescape/console/repository/MemoryReservationTimeRepository.java
@@ -1,9 +1,11 @@
-package roomescape.domain;
+package roomescape.console.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicLong;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 
 public class MemoryReservationTimeRepository implements ReservationTimeRepository {
