+++ src/main/java/roomescape/domain/ReservationService.java
@@ -2,8 +2,8 @@ package roomescape.domain;
 
 import java.util.List;
 import org.springframework.stereotype.Service;
-import roomescape.controller.dto.ReservationRequest;
-import roomescape.controller.dto.ReservationResponse;
+import roomescape.domain.dto.ReservationRequest;
+import roomescape.domain.dto.ReservationResponse;
 
 @Service
 public class ReservationService {
