+++ src/main/java/roomescape/domain/dto/ReservationTimeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.domain.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalTime;
