+++ src/main/java/roomescape/web/controller/dto/ErrorResponse.java
@@ -1,4 +1,4 @@
-package roomescape.controller.dto;
+package roomescape.web.controller.dto;
 
 public record ErrorResponse(
         String message) {
